remove cluster option
