+++ package.json
@@ -44,13 +44,12 @@
     "leaflet": "^1.9.4",
     "lodash": "^4.17.21",
     "notistack": "^3.0.1",
-    "react": "^18.3.1",
-    "react-dom": "^18.3.1",
+    "react": "^19.0.0",
+    "react-dom": "^19.0.0",
     "react-error-boundary": "^5.0.0",
     "react-hook-form": "^7.54.0",
     "react-i18next": "^15.1.4",
-    "react-leaflet": "^4.2.1",
-    "react-leaflet-cluster": "^2.1.0"
+    "react-leaflet": "^5.0.0"
   },
   "devDependencies": {
     "@casl/react": "^5.0.0",
@@ -75,8 +74,8 @@
     "@types/leaflet": "^1.9.16",
     "@types/lodash": "^4.17.13",
     "@types/node": "^22.10.2",
-    "@types/react": "^18.3.18",
-    "@types/react-dom": "^18.3.5",
+    "@types/react": "^19.0.8",
+    "@types/react-dom": "^19.0.3",
     "@vitejs/plugin-react": "^4.3.4",
     "date-fns": "^4.1.0",
     "eslint": "^9.15.0",
@@ -89,18 +88,17 @@
     "lodash": "^4.17.21",
     "notistack": "^3.0.1",
     "prettier": "^3.4.2",
-    "react": "^18.3.1",
-    "react-dom": "^18.3.1",
+    "react": "^19.0.0",
+    "react-dom": "^19.0.0",
     "react-error-boundary": "^5.0.0",
     "react-hook-form": "^7.54.0",
     "react-i18next": "^15.1.4",
-    "react-leaflet": "^4.2.1",
+    "react-leaflet": "^5.0.0",
     "storybook": "^8.5.2",
     "typescript": "~5.7.2",
     "typescript-eslint": "^8.15.0",
     "vite": "^6.0.1",
-    "vite-plugin-dts": "^4.3.0",
-    "react-leaflet-cluster": "^2.1.0"
+    "vite-plugin-dts": "^4.3.0"
   },
   "eslintConfig": {
     "extends": [
