+++ vite.config.ts
@@ -24,7 +24,6 @@ export default defineConfig({
         'react',
         'react-dom',
         'react/jsx-runtime',
-        'react-leaflet-cluster',
       ],
     },
   },
